+++ apps/mobile-app/e2e/tests/playgroundRoutes.e2e.ts
@@ -14,6 +14,10 @@ const disabledRoutes = {
     'Tray', // pointless
     'LottieStatusAnimation' /** Animation */,
     'Modal' /** Modal is displayed over expected screen */,
+    'PatternTagDisclosureHighFrictionBenefit' /** Modal is displayed over expected screen, and also Android specific issue preventing navigation to screen */,
+    'PatternTagDisclosureHighFrictionRisk' /** Modal is displayed over expected screen, and also Android specific issue preventing navigation to screen */,
+    'PatternTagDisclosureLowFriction' /** Modal is displayed over expected screen, and also Android specific issue preventing navigation to screen */,
+    'PatternTagDisclosureMedFriction' /** Modal is displayed over expected screen, and also Android specific issue preventing navigation to screen */,
     'PatternTagError' /** Modal is displayed over expected screen, and also Android specific issue preventing navigation to screen */,
   ],
   iosDisabledRoutes: [],